# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The",
        "title": "MuXiFresh-be",
        "contact": {},
        "version": "1.0"
    },
    "host": "work.test.muxi-tech.xyz",
    "basePath": "/api/v1",
    "paths": {
...
        "/form": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "form"
                ],
                "summary": "\"编辑报名表\"",
                "parameters": [
                    {
                        "description": "edit_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/editRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/form.FormModel"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "form"
                ],
                "summary": "\"创建报名表\"",
                "parameters": [
                    {
                        "description": "create_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"创建报名表成功\"}",
                        "schema": {
                            "$ref": "#/definitions/form.FormModel"
                        }
                    }
                }
            }
        },
        "/form/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "form"
                ],
                "summary": "\"查看报名表（分组）\"",
                "parameters": [
                    {
                        "description": "按组查询报名表",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.searchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/homework/change/published/:id": {
            "post": {
                "description": "修改已发布的作业",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Modify published homework",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改内容",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/homework.ModifyPublishedHomeworkRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "发布的的作业id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    }
                }
            }
        },
...
        "/homework/comment": {
            "get": {
                "description": "查看已发布的评论",
                "consumes": [
                    "json/application"
                ],
                "produces": [
                    "json/application"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Get comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "获取email",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id--帖子的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit--偏移量指定开始返回记录之前要跳过的记录数 ",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page--限制指定要检索的记录数 ",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Add comment to one's homework",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Comment on Homework",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "HomeworkID 评论作业的id || Content 评论内容",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/homework.CommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    }
                }
            }
        },
        "/homework/comment/:comment_id": {
            "delete": {
                "description": "删除用户发布的帖子",
                "consumes": [
                    "json/application"
                ],
                "produces": [
                    "json/application"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Delete comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "获取email",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "评论的id",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    }
                }
            }
        },
...
        "/homework/publish": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Publish homework",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "content--作业内容 ; group_id--小组id   1-设计组 2-产品组 3-安卓组 4-前端组 5-后端组",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/homework.HomeworkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errno.Errno"
                        }
                    }
                }
            }
        },
...
        "/homework/published/details/:id": {
            "get": {
                "description": "get the published homework details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "Get homework details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "发布的作业id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"查看成功\"}",
                        "schema": {
                            "$ref": "#/definitions/file.HomeworkPublished"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "createRequest": {
            "type": "object",
            "required": [
                "avatar",
                "college",
                "contact_number",
                "contact_way",
                "gender",
                "grade",
                "group",
                "if_other_organization",
                "major",
                "name",
                "reason",
                "self_introduction",
                "student_id",
                "understand"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "type": "string"
                },
                "contact_number": {
                    "type": "string"
                },
                "contact_way": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "if_other_organization": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "self_introduction": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "understand": {
                    "type": "string"
                }
            }
        },
        "editRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "type": "string"
                },
                "contact_number": {
                    "type": "string"
                },
                "contact_way": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "if_other_organization": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "self_introduction": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "understand": {
                    "type": "string"
                }
            }
        },
        "errno.Errno": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "file.Homework": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "homework_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "file.HomeworkPublished": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "homeworks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/file.Homework"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "form.FormModel": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "type": "string"
                },
                "contact_number": {
                    "type": "string"
                },
                "contact_way": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "if_other_organization": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "self_introduction": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "understand": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "form.searchRequest": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                }
            }
        },
        "homework.CommentRequest": {
            "type": "object",
            "required": [
                "content",
                "homework_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "homework_id": {
                    "type": "integer"
                }
            }
        },
        "homework.HomeworkRequest": {
            "type": "object",
            "required": [
                "content",
                "homework_id",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "homework_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "homework.ModifyPublishedHomeworkRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
...
}